Move Info Menu to its own file
